[refactor] 생년월일을 통신 파라미터에 넣어 보낼 때, Int 로의 형변환을 repository 에서 수행하도록 변경

+++ app/src/main/java/com/jangjh123/allpouse_android/ui/screen/login/LoginActivity.kt
@@ -307,7 +307,7 @@ class LoginActivity : ComponentActivity() {
                                                 socialId = socialId,
                                                 userName = nicknameState.value,
                                                 permission = "ROLE_USER",
-                                                age = ageState.value.toInt(),
+                                                age = ageState.value,
                                                 gender = when (genderState.value) {
                                                     Man -> {
                                                         "Man"
